# openapi.yaml
openapi: 3.0.0
info:
  title: Authorization Microservice
  version: 1.0.0
  description: Authorization microservice for medical consultation application. Handles the authorization of users for the entire application.
servers:
  - url: http://localhost:3001/auth/api/v1
paths:
...
  /validate:
    get:
      summary: Validates the user's token.
      description: Checks the validity of the token provided in the user's cookies.
      tags:
        - Authentication
      operationId: validateToken
      responses:
        '200':
          description: Token is valid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token is valid
        '401':
          description: Unauthorized - Token is missing, expired, or invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token expired or invalid.
components:
...
  securitySchemes:
    cookieAuth:
        type: apiKey
        in: cookie
        name: token
    # Define any security schemes if relevant (e.g., JWT authentication).

  # Add other reusable components here (e.g., responses, request bodies).
...